server-side caching implemented

+++ src/services/notes/repositories/note-repositories.js
@@ -1,11 +1,13 @@
 import { nanoid } from 'nanoid';
 import { Pool } from 'pg';
+import CacheService from '../../../cache/redis-service.js';
 import collaborationRepositories from '../../collaborations/repositories/collaboration-repositories.js';
 
 class NoteRepositories {
   constructor() {
     this.pool = new Pool();
     this.collaborationRepositories = collaborationRepositories;
+    this.cacheService = new CacheService();
   }
 
   async createNote({ title, body, tags, owner }) {
@@ -19,21 +21,38 @@ class NoteRepositories {
     };
 
     const result = await this.pool.query(query);
+
+    // delete cache
+    await this.cacheService.delete(`notes:${owner}`);
+
     return result.rows[0];
   }
 
   async getNotes(owner) {
-    const query = {
-      text: `SELECT notes.* 
+    const cacheKey = `notes:${owner}`;
+
+    try {
+      const notes = await this.cacheService.get(cacheKey);
+      return JSON.parse(notes);
+      // eslint-disable-next-line no-unused-vars
+    } catch (error) {
+      // Cache miss, get from database
+      const query = {
+        text: `SELECT notes.* 
               FROM notes
               LEFT JOIN collaborations ON collaborations.note_id = notes.id
               WHERE notes.owner = $1 OR collaborations.user_id = $1
               GROUP BY notes.id`,
-      values: [owner],
-    };
+        values: [owner],
+      };
 
-    const result = await this.pool.query(query);
-    return result.rows;
+      const result = await this.pool.query(query);
+
+      // save to cache
+      await this.cacheService.set(cacheKey, JSON.stringify(result.rows));
+
+      return result.rows;
+    }
   }
 
   async getNoteById(id) {
@@ -54,21 +73,33 @@ class NoteRepositories {
     const updatedAt = new Date().toISOString();
 
     const query = {
-      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id',
+      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner',
       values: [title, body, tags, updatedAt, id],
     };
 
     const result = await this.pool.query(query);
+
+    const owner = result.rows[0].owner;
+    if (result.rows[0]) {
+      await this.cacheService.delete(`notes:${owner}`);
+    }
+
     return result.rows[0];
   }
 
   async deleteNote(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
     const result = await this.pool.query(query);
+
+    const owner = result.rows[0].owner;
+
+    if (result.rows[0]) {
+      await this.cacheService.delete(`notes:${owner}`);
+    }
 
     return result.rows[0].id;
   }
